+++ .env.example
@@ -7,3 +7,6 @@ export EMAIL_HOST_USER=<>
 export EMAIL_HOST_PASSWORD=<>
 export DEFAULT_FROM_EMAIL=<>
 export EMAIL_HOST=<>
+
+export SECRET_KEY=<>
+export DEBUG=<>
