+++ .github/workflows/deploy.yaml
@@ -1,0 +1,27 @@
+# name of our workflow
+name: Deployment
+
+on:
+  # any code pushed to master branch will trigger workflow
+  push:
+    branches: [ master ]
+
+jobs:
+  Deploy: # deploy job for deploying code to heroku
+    runs-on: ubuntu-latest # os for running the job
+
+    steps:
+    - uses: actions/checkout@v2
+    - run: |
+        git fetch --prune --unshallow
+    - name: Deploy to Heroku
+      env:
+        HEROKU_API_TOKEN: ${{ secrets.HEROKU_API_TOKEN }} # heroku api token for deployment
+        HEROKU_APP_NAME: ${{ secrets.HEROKU_APP_NAME }} # heroku app name
+      if: github.ref == 'refs/heads/master' && job.status == 'success'
+      run: |
+        git remote add heroku https://heroku:$HEROKU_API_TOKEN@git.heroku.com/$HEROKU_APP_NAME.git
+        git push heroku HEAD:master -f
+
+    - name: Successful Message # success message
+      run: echo Successfully Deployed!!
