--- ExpenseTracker/ExpenseTracker/static/js/main.js
@@ -1,1 +1,0 @@
-console.log("1111", 1111);
