+++ requirements.txt
@@ -2,7 +2,7 @@ asgiref==3.3.4
 coverage==5.5
 Django==3.2
 gunicorn==20.1.0
-psycopg2==2.7.7
+psycopg2==2.8.4
 psycopg2-binary==2.8.6
 pycodestyle==2.7.0
 pytz==2021.1
