+++ templates/income/add_income.html
